refactor(scalars): add OIDInput & refactor IdAutocomplete base component
