+++ packages/design-system/src/scalars/components/aid-field/types.ts
@@ -1,7 +1,7 @@
 import type {
   IdAutocompleteOption,
   IdAutocompleteProps,
-} from "../fragments/id-autocomplete-field/types.js";
+} from "../fragments/id-autocomplete/types.js";
 
 export interface Network {
   chainId: string;
