--- packages/design-system/src/scalars/components/fragments/id-autocomplete-field/index.ts
@@ -1,1 +1,0 @@
-export * from "./id-autocomplete-field.js";
