+++ packages/design-system/src/scalars/components/fragments/id-autocomplete/index.ts
@@ -1,0 +1,2 @@
+export { IdAutocomplete } from "./id-autocomplete.js";
+export type { IdAutocompleteOption, IdAutocompleteProps } from "./types.js";
