+++ packages/design-system/src/scalars/components/oid-field/index.ts
@@ -1,1 +1,2 @@
-export * from "./oid-field.js";
+export { OIDField } from "./oid-field.js";
+export type { OIDFieldProps, OIDOption } from "./types.js";
