+++ packages/design-system/src/scalars/components/phid-field/mocks.ts
@@ -1,4 +1,4 @@
-import type { IdAutocompleteOption } from "../fragments/id-autocomplete-field/types.js";
+import type { IdAutocompleteOption } from "../fragments/id-autocomplete/types.js";
 
 export const mockedOptions: IdAutocompleteOption[] = [
   {
